+++ app/src/main/java/com/example/mapYandex/TagDao.kt
@@ -13,13 +13,14 @@ interface TagDao {
     suspend fun insert(vararg tags: Tag)
 
     @Insert
-    suspend fun insert(tag: Tag): Int
+    suspend fun insert(tag: Tag): Long
 
     @Query("SELECT * FROM tag")
     fun findAll(): LiveData<List<Tag>>
 
     @Query("SELECT * FROM tag WHERE id=:id LIMIT 1")
-    fun findById(id: Int): LiveData<Tag>
+    fun findById(id: Long): LiveData<Tag>
+
 
     @Update
     suspend fun update(tag: Tag): Int
